project/inference-api/requirements.txt
fastapi==0.95.0
uvicorn[standard]==0.22.0
onnxruntime==1.15.0
Pillow==10.0.0
numpy==1.24.2
matplotlib==3.7.1
torch==2.0.0
torchvision==0.15.0
scikit-learn==1.2.2
opencv-python==4.8.0.76
tqdm==4.65.0
